// frontend/src/app/model/Engagement.ts
export class Engagement {
  engag_numeroengagement: Number;
  engag_numpoli: Number;
  engag_numeroavenant: Number;
  engag_numeroacte: Number;
  engag_codemarche: Number;
  engag_kapassure: number;
  engag_dateengagement: Date;
  engag_capitalliberationengage: Number;
  engag_dateliberation: Date;
  engag_typesurete: String;
  engag_identificationtitre: String;
  engag_retenudeposit: Number;
  engag_datedeposit: Date;
  engag_depositlibere: Number;
  engag_dateliberationdeposit: Date;
  engag_cautionsolidaire: String;
  engag_datecomptabilisation: Date;
  engag_status: String;
  engag_codeutilisateur: String;
  engag_datemodification: Date;
  active: Number;
}